Extracted Subscription Welcome Email into its own component
